Appease pylint, mypy, and pytest
