+++ tests/context.py
@@ -9,3 +9,4 @@ import amplipi.hw
 import amplipi.models
 import amplipi.rt # TODO: remove
 import amplipi.streams
+import amplipi.utils
